feat: adapting libc_trusty
Started addapting libc_trusty. Main goal is to implement a model for ipc.c

First step is an implementation of the handle table.

This is based on earlier implementation by Xiang. The goal is to replace
that implementation with a more flexible solution that allows for multiple
ports and channels.