fix: enable user to provide c++ stdlib to use
On some machines libc++-10-dev is too old and a newer version libc++ is available which is not picked up during compile when clang version is set to 10. Forcing the gnu stdlib to be picked up fixes missing headers. Note that the default is to use libc++ since the CI OS is old enough to have libc++-10-dev.